# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookal venue booking management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  fullName: string;
  email: string;
  role: string;
  mustChangePw: boolean;
}

export interface LoginResponse {
  token: string;
  refreshToken?: string;
  user: AuthUser;
}

export interface ChangePasswordRequest {
  currentPassword?: string;
  newPassword: string;
}

export interface User {
  id: string;
  fullName: string;
  email: string;
  role: string;
  isActive: boolean;
  mustChangePw: boolean;
  createdAt: string;
  lastLogin?: string | null;
}

export interface CreateUserRequest {
  fullName: string;
  email: string;
  password: string;
  role: string;
}

export interface UpdateUserRequest {
  fullName?: string;
  role?: string;
  isActive?: boolean;
}

export interface Venue {
  id: string;
  name: string;
  type: string;
  pricePerHour: number;
  isActive: boolean;
  displayOrder: number;
}

export interface BookingVenueItem {
  venueId: string;
  pricePerHour: number;
}

export interface BookingVenueDetail {
  id: string;
  venueId: string;
  venueName: string;
  venueType: string;
  pricePerHour: number;
  subtotal: number;
}

export interface Booking {
  id: string;
  bookingRef: string;
  customerName: string;
  phoneNumbers: string[];
  address?: string | null;
  idProofUrl?: string | null;
  bookingDate: string;
  tamilDateLabel?: string | null;
  startTime: string;
  endTime: string;
  durationHours: number;
  totalAmount: number;
  advanceAmount?: number;
  isPaid?: boolean;
  notes?: string | null;
  status: string;
  createdById: string;
  createdByName: string;
  createdAt: string;
  updatedAt: string;
  cancelReason?: string | null;
  venues: BookingVenueDetail[];
}

export type BookingDetail = Booking;

export interface CreateBookingRequest {
  customerName: string;
  phoneNumbers: string[];
  address?: string | null;
  idProofUrl?: string | null;
  bookingDate: string;
  tamilDateLabel?: string | null;
  startTime: string;
  endTime: string;
  venues: BookingVenueItem[];
  advanceAmount?: number;
  isPaid?: boolean;
  notes?: string | null;
}

export interface BookingsListResponse {
  bookings: Booking[];
  total: number;
  page: number;
  totalPages: number;
}

export type AvailabilityResponseConflict = {
  bookingRef: string;
  customerName: string;
  startTime: string;
  endTime: string;
} | null;

export interface AvailabilityResponse {
  available: boolean;
  conflict?: AvailabilityResponseConflict;
}

export type ConflictResponseConflictsItem = {
  venueName: string;
  bookingRef: string;
  customerName: string;
  startTime: string;
  endTime: string;
};

export interface ConflictResponse {
  error: string;
  conflicts?: ConflictResponseConflictsItem[];
}

export interface SettingsResponse {
  default_duration_hours: string;
  session_timeout_hours: string;
  rules_pdf_path: string;
  biz_name?: string | null;
  biz_tagline?: string | null;
  biz_address?: string | null;
  biz_phone?: string | null;
  biz_email?: string | null;
  biz_gst?: string | null;
}

export type ReportSummaryByVenueItem = {
  venueId: string;
  venueName: string;
  bookingCount: number;
  revenue: number;
};

export type ReportSummaryByDayItem = {
  date: string;
  bookingCount: number;
  revenue: number;
};

export type ReportSummaryByEmployeeItem = {
  userId: string;
  userName: string;
  bookingCount: number;
  revenue: number;
};

export interface ReportSummary {
  totalBookings: number;
  totalRevenue: number;
  confirmedBookings: number;
  cancelledBookings: number;
  avgBookingValue: number;
  byVenue: ReportSummaryByVenueItem[];
  byDay: ReportSummaryByDayItem[];
  byEmployee: ReportSummaryByEmployeeItem[];
}

export interface CustomerSearchResult {
  customerName: string;
  phoneNumbers: string[];
  address?: string | null;
}

export type GetBookingsParams = {
  date?: string;
  status?: string;
  search?: string;
  from?: string;
  to?: string;
  page?: number;
  limit?: number;
};

export type CheckAvailabilityParams = {
  venueId: string;
  date: string;
  startTime: string;
  endTime: string;
  excludeBookingId?: string;
};

export type CancelBookingBody = {
  reason: string;
};

export type UpdateVenuePriceBody = {
  pricePerHour: number;
};

export type UploadRulesPdfBody = {
  pdfData: string;
};

export type GetReportSummaryParams = {
  from: string;
  to: string;
};

export type ExportReportParams = {
  from: string;
  to: string;
};

export type SearchCustomersParams = {
  q: string;
};