# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bookal venue booking management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get report summary (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report as CSV (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - token
        - user
    AuthUser:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        mustChangePw:
          type: boolean
      required:
        - id
        - fullName
        - email
        - role
        - mustChangePw
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - newPassword
...
    ReportSummary:
      type: object
      properties:
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        confirmedBookings:
          type: integer
        cancelledBookings:
          type: integer
        avgBookingValue:
          type: number
        byVenue:
          type: array
          items:
            type: object
            properties:
              venueId:
                type: string
              venueName:
                type: string
              bookingCount:
                type: integer
              revenue:
                type: number
            required:
              - venueId
              - venueName
              - bookingCount
              - revenue
        byDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              bookingCount:
                type: integer
              revenue:
                type: number
            required:
              - date
              - bookingCount
              - revenue
        byEmployee:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              userName:
                type: string
              bookingCount:
                type: integer
              revenue:
                type: number
            required:
              - userId
              - userName
              - bookingCount
              - revenue
      required:
        - totalBookings
        - totalRevenue
        - confirmedBookings
        - cancelledBookings
        - avgBookingValue
        - byVenue
        - byDay
        - byEmployee
...